+++ demos/demo/lib/screens/home_screen.dart
@@ -1,7 +1,9 @@
 import 'package:flutter/material.dart';
 
+import 'lx_computed_demo.dart';
 import 'lx_data_builder_demo.dart';
 import 'lx_multi_builder_demo.dart';
+import 'lx_operators_demo.dart';
 import 'lx_state_builder_demo.dart';
 import 'lx_view_demo.dart';
 import 'lx_widget_demo.dart';
@@ -56,6 +58,22 @@ class HomeScreen extends StatelessWidget {
               onPressed: () {
                 Navigator.of(context)
                     .push(MaterialPageRoute(builder: (_) => LxMultiDemoView()));
+              },
+            ),
+            // Stream Operators
+            ElevatedButton(
+              child: Text('Stream Operators Demo'),
+              onPressed: () {
+                Navigator.of(context).push(MaterialPageRoute(
+                    builder: (_) => LxOperatorsDemoView()));
+              },
+            ),
+            // Computed Values
+            ElevatedButton(
+              child: Text('Computed Values Demo'),
+              onPressed: () {
+                Navigator.of(context).push(MaterialPageRoute(
+                    builder: (_) => LxComputedDemoView()));
               },
             ),
           ],